OCT-111: moved GetMeasurementsFamilyQuery
