+++ components/catalogs/back/src/Infrastructure/Symfony/Resources/config/services.yml
@@ -132,7 +132,7 @@ services:
         arguments:
             $getConnectorProducts: '@akeneo.pim.enrichment.product.connector.get_product_from_uuids'
 
-    Akeneo\Catalogs\Infrastructure\Persistence\GetMeasurementsFamilyQuery:
+    Akeneo\Catalogs\Infrastructure\Persistence\Measurement\GetMeasurementsFamilyQuery:
         arguments:
             - '@akeneo_measure.persistence.measurement_family_repository'
