+++ components/catalogs/back/tests/Integration/Infrastructure/Validation/ProductSelection/AttributeCriterion/AbstractAttributeCriterionTest.php
@@ -9,13 +9,13 @@ use Akeneo\Catalogs\Application\Persistence\GetAttributeOptionsByCodeQueryInterf
 use Akeneo\Catalogs\Application\Persistence\GetChannelLocalesQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\GetChannelQueryInterface;
 use Akeneo\Catalogs\Application\Persistence\GetLocalesQueryInterface;
-use Akeneo\Catalogs\Application\Persistence\GetMeasurementsFamilyQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Measurement\GetMeasurementsFamilyQueryInterface;
 use Akeneo\Catalogs\Infrastructure\Persistence\FindOneAttributeByCodeQuery;
 use Akeneo\Catalogs\Infrastructure\Persistence\GetAttributeOptionsByCodeQuery;
 use Akeneo\Catalogs\Infrastructure\Persistence\GetChannelLocalesQuery;
 use Akeneo\Catalogs\Infrastructure\Persistence\GetChannelQuery;
 use Akeneo\Catalogs\Infrastructure\Persistence\GetLocalesQuery;
-use Akeneo\Catalogs\Infrastructure\Persistence\GetMeasurementsFamilyQuery;
+use Akeneo\Catalogs\Infrastructure\Persistence\Measurement\GetMeasurementsFamilyQuery;
 use Akeneo\Catalogs\Test\Integration\IntegrationTestCase;
 
 /**
